+++ Engine/TMSoundEngine.mm
@@ -73,7 +73,7 @@ void *getOpenALAudioData(CFURLRef inFileURL, ALsizei *outDataSize, ALenum *outDa
     err = AudioFileOpenURL(inFileURL, kAudioFileReadPermission, 0, &aFID);
     if ( err )
     {
-        printf("getOpenALAudioData: AudioFileOpenURL FAILED, Error = %ld\n", err);
+        printf("getOpenALAudioData: AudioFileOpenURL FAILED, Error = %d\n", (int)err);
         goto Exit;
     }
 
@@ -81,7 +81,7 @@ void *getOpenALAudioData(CFURLRef inFileURL, ALsizei *outDataSize, ALenum *outDa
     err = AudioFileGetProperty(aFID, kAudioFilePropertyDataFormat, &thePropertySize, &theFileFormat);
     if ( err )
     {
-        printf("getOpenALAudioData: AudioFileGetProperty(kAudioFilePropertyDataFormat) FAILED, Error = %ld\n", err);
+        printf("getOpenALAudioData: AudioFileGetProperty(kAudioFilePropertyDataFormat) FAILED, Error = %d\n", (int)err);
         goto Exit;
     }
     if ( theFileFormat.mChannelsPerFrame > 2 )
@@ -107,7 +107,7 @@ void *getOpenALAudioData(CFURLRef inFileURL, ALsizei *outDataSize, ALenum *outDa
     err = AudioFileGetProperty(aFID, kAudioFilePropertyAudioDataByteCount, &thePropertySize, &theFileLengthInBytes);
     if ( err )
     {
-        printf("getOpenALAudioData: AudioFileGetProperty(kAudioFilePropertyAudioDataByteCount) FAILED, Error = %ld\n", err);
+        printf("getOpenALAudioData: AudioFileGetProperty(kAudioFilePropertyAudioDataByteCount) FAILED, Error = %d\n", (int)err);
         goto Exit;
     }
 
@@ -137,7 +137,7 @@ void *getOpenALAudioData(CFURLRef inFileURL, ALsizei *outDataSize, ALenum *outDa
             // failure
             free(theData);
             theData = NULL; // make sure to return NULL
-            printf("getOpenALAudioData: AudioFileRead FAILED, Error = %ld\n", err);
+            printf("getOpenALAudioData: AudioFileRead FAILED, Error = %d\n", (int)err);
             goto Exit;
         }
     }
